Refactor: Extract core logic into modular packages and modules
