+++ package/domain/models.py
@@ -1,0 +1,13 @@
+from enum import IntEnum, Enum
+
+
+class Drafted(IntEnum):
+    NONE = 0
+    OTHER = 1
+    MINE = 2
+
+
+class Sport(str, Enum):
+    NFL = "nfl"
+    NBA = "nba"
+
