+++ package/mainwindow.py
@@ -32,6 +32,9 @@ from sklearn.linear_model import Lasso, LassoLarsIC
 from sklearn.tree import DecisionTreeRegressor
 
 from package.ui.mainwindow_ui import Ui_Form
+from package.optimization.team_optimizer import find_optimal_team_sat, create_team_df
+from package.analysis.opportunity_cost import calc_player_oc
+from package.data.repositories import save_draft_state, load_most_recent_draft_state
 
 
 class QCustomTableWidgetItem(qtw.QTableWidgetItem):
@@ -359,19 +362,13 @@ class MainWindow(qtw.QWidget):  # Would be something else if you didn't use widg
         # print(opt_team.loc[opt_team["Position"] != "B", "PPW"].sum())
 
     def export_draft(self):
-        self.draft_df.to_csv(
-            f".cache/draft_states/{datetime.datetime.now().strftime('%Y_%m_%d_%H_%M_%S')}.csv"
-        )
+        save_draft_state(self.draft_df)
 
     def import_draft(self):
-        base_path = Path(".cache/draft_states")
-        files = [x for x in base_path.glob("*.csv") if x.is_file()]
-        dates = []
-        for file in files:
-            date = datetime.datetime.strptime(file.stem, "%Y_%m_%d_%H_%M_%S")
-            dates.append(date)
-        most_recent_draft_ind = np.argmax(dates)
-        self.draft_df = pd.read_csv(files[most_recent_draft_ind])
+        df = load_most_recent_draft_state()
+        if df is None:
+            return
+        self.draft_df = df
         self.draft_df = self.draft_df.fillna({"T/A": ""})
         self.calc_player_price()
 
@@ -2136,44 +2133,7 @@ class MainWindow(qtw.QWidget):  # Would be something else if you didn't use widg
         self.oc_canvas.draw()
 
 
-def calc_player_oc(
-    player_name,
-    test_price,
-    draft_df,
-    league_roster,
-    auction_budget,
-    starter_percent,
-    team_cols,
-    sport,
-):
-    player_ind = draft_df.loc[draft_df["Name"] == player_name].index
-
-    spent_dollars = draft_df.loc[draft_df["Drafted"] == 2, "Draft$"].sum()
-
-    # Get the number of starters remaining at the position
-    player_position = draft_df.loc[player_ind, "Position"].values[0]
-
-    test_df = copy.deepcopy(draft_df)
-    # Get the optimal team if the player is drafted at the test price
-    test_df.loc[player_ind, "Drafted"] = 2
-    test_df.loc[player_ind, "Draft$"] = test_price
-    opt_take_team = find_optimal_team_sat(
-        test_df, league_roster, auction_budget, starter_percent, team_cols, sport
-    )
-    if opt_take_team is None:
-        return 100
-    opt_take_ppw = opt_take_team["PPW"].sum()
-
-    test_df.loc[player_ind, "Drafted"] = 1
-    test_df.loc[player_ind, "Draft$"] = test_price
-    opt_leave_team = find_optimal_team_sat(
-        test_df, league_roster, auction_budget, starter_percent, team_cols, sport
-    )
-    if opt_leave_team is None:
-        return 100
-    opt_leave_ppw = opt_leave_team["PPW"].sum()
-    opportunity_cost = opt_leave_ppw - opt_take_ppw
-    return opportunity_cost
+# Moved to package.analysis.opportunity_cost.calc_player_oc
 
 
 # def find_optimal_team(draft_state_df, league_roster, my_starter_budget, team_cols):
@@ -2249,462 +2209,7 @@ def calc_player_oc(
 #     return opt_team
 
 
-def find_optimal_team_sat(
-    draft_state_df, league_roster, auction_budget, starter_percent, team_cols, sport
-):
-    starter_budget = int(auction_budget * starter_percent)
-
-    model = cp_model.CpModel()
-
-    # Replace draft board with only undrafted or my players
-    draft_state_df = draft_state_df.loc[draft_state_df["Drafted"] != 1].reset_index(
-        drop=True
-    )
-
-    if sport == "nfl":
-        # Set up starter/bench booleans
-        starter_bools = []
-        bench_bools = []
-        for i, player in draft_state_df.iterrows():
-            starter_bool = model.NewBoolVar(f"{player.Name} is starter")
-            starter_bools.append(starter_bool)
-
-            bench_bool = model.NewBoolVar(f"{player.Name} is on bench")
-            bench_bools.append(bench_bool)
-
-            model.Add(starter_bool + bench_bool <= 1)
-
-        n_team = (
-            league_roster["QB"]
-            + league_roster["RB"]
-            + league_roster["WR"]
-            + league_roster["TE"]
-            + league_roster["FLEX"]
-            + league_roster["B"]
-        )
-        model.Add(sum(starter_bools) + sum(bench_bools) == n_team)
-        # Positional constraints
-        nf = league_roster["FLEX"]
-        n_flex = (
-            league_roster["RB"]
-            + league_roster["WR"]
-            + league_roster["TE"]
-            + league_roster["FLEX"]
-        )
-        n_drafted_qb = draft_state_df.loc[
-            (draft_state_df["Drafted"] == 2) & (draft_state_df["Position"] == "QB")
-        ].shape[0]
-        n_drafted_rb = draft_state_df.loc[
-            (draft_state_df["Drafted"] == 2) & (draft_state_df["Position"] == "RB")
-        ].shape[0]
-        n_drafted_wr = draft_state_df.loc[
-            (draft_state_df["Drafted"] == 2) & (draft_state_df["Position"] == "WR")
-        ].shape[0]
-        n_drafted_te = draft_state_df.loc[
-            (draft_state_df["Drafted"] == 2) & (draft_state_df["Position"] == "TE")
-        ].shape[0]
-        n_drafted_flex = draft_state_df.loc[
-            (draft_state_df["Drafted"] == 2)
-            & (draft_state_df["Position"].isin(["RB", "WR", "TE"]))
-        ].shape[0]
-
-        qb_target_terms = []
-        rb_target_terms = []
-        wr_target_terms = []
-        te_target_terms = []
-        flex_target_terms = []
-        bench_target_terms = []
-        qb_terms = []
-        rb_terms = []
-        wr_terms = []
-        te_terms = []
-        flex_terms = []
-        bench_terms = []
-        for i, player in draft_state_df.iterrows():
-            pos = player["Position"]
-            if (player["T/A"] == "T") & (player["Drafted"] != 2):
-                qb_target_terms.append(starter_bools[i] * int(pos == "QB"))
-                rb_target_terms.append(starter_bools[i] * int(pos == "RB"))
-                wr_target_terms.append(starter_bools[i] * int(pos == "WR"))
-                te_target_terms.append(starter_bools[i] * int(pos == "TE"))
-                flex_target_terms.append(
-                    starter_bools[i] * int(pos in ["RB", "WR", "TE"])
-                )
-                bench_target_terms.append(bench_bools[i] * int(pos in ["RB", "WR"]))
-            qb_terms.append(starter_bools[i] * int(pos == "QB"))
-            rb_terms.append(starter_bools[i] * int(pos == "RB"))
-            wr_terms.append(starter_bools[i] * int(pos == "WR"))
-            te_terms.append(starter_bools[i] * int(pos == "TE"))
-            flex_terms.append(starter_bools[i] * int(pos in ["RB", "WR", "TE"]))
-            bench_terms.append(bench_bools[i] * int(pos in ["RB", "WR"]))
-
-        model.Add(sum(bench_terms) == league_roster["B"])
-        model.Add(sum(qb_terms) == league_roster["QB"])
-
-        model.Add(sum(rb_terms) >= league_roster["RB"])
-        model.Add(sum(rb_terms) <= league_roster["RB"] + nf)
-
-        model.Add(sum(wr_terms) >= league_roster["WR"])
-        model.Add(sum(wr_terms) <= league_roster["WR"] + nf)
-
-        model.Add(sum(te_terms) >= league_roster["TE"])
-        model.Add(sum(te_terms) <= league_roster["TE"] + nf)
-
-        model.Add(sum(flex_terms) == n_flex)
-
-        model.Add(sum(bench_terms) == league_roster["B"])
-
-        # On team and price constraints
-
-        # already_on_team = []
-        starter_price_terms = []
-        bench_price_terms = []
-        for i, player in draft_state_df.iterrows():
-            if player["Drafted"] == 2:
-                starter_price_terms.append(
-                    starter_bools[i] * int(round(player["Draft$"], 0))
-                )
-                bench_price_terms.append(
-                    bench_bools[i] * int(round(player["Draft$"], 0))
-                )
-                model.Add(starter_bools[i] + bench_bools[i] == 1)
-            elif player["T/A"] == "A":
-                model.Add(starter_bools[i] + bench_bools[i] == 0)
-            else:
-                starter_price_terms.append(
-                    starter_bools[i] * int(round(player["Proj$"], 0))
-                )
-                bench_price_terms.append(
-                    bench_bools[i] * int(round(player["Proj$"], 0))
-                )
-        model.Add(sum(starter_price_terms) <= starter_budget)
-        model.Add(sum(starter_price_terms) + sum(bench_price_terms) <= auction_budget)
-        # model.Add(sum(already_on_team) == len(already_on_team))
-
-        starter_values = []
-        bench_values = []
-        for i, player in draft_state_df.iterrows():
-            starter_values.append(starter_bools[i] * int(1000 * player.PPW))
-            bench_values.append(bench_bools[i] * int(1000 * player.PPW))
-        model.Maximize(sum(starter_values) + 0.01 * sum(bench_values))
-        solver = cp_model.CpSolver()
-        # solver.parameters.num_search_workers = 12
-        # solver.parameters.log_search_progress = True
-        # solver.parameters.max_time_in_seconds = self.obs_scheme[
-        #     "max_time_in_seconds"
-        # ]
-        status = solver.Solve(model)
-        if status == cp_model.OPTIMAL or status == cp_model.FEASIBLE:
-            starters = []
-            bench = []
-            for i, player in draft_state_df.iterrows():
-                if solver.Value(starter_bools[i]):
-                    starters.append(player["Name"])
-                elif solver.Value(bench_bools[i]):
-                    bench.append(player["Name"])
-            starters.extend(bench)
-            opt_team = create_team_df(
-                starters, league_roster, draft_state_df, team_cols
-            )
-        else:
-            opt_team = None
-    elif sport == "nba":
-        n_team = (
-            league_roster["PG"]
-            + league_roster["SG"]
-            + league_roster["SF"]
-            + league_roster["PF"]
-            + league_roster["C"]
-            + league_roster["G"]
-            + league_roster["F"]
-            + league_roster["UTIL"]
-            + league_roster["B"]
-        )
-        player_vars = []
-        for i, player in draft_state_df.iterrows():
-            player_var = model.NewBoolVar(f"{player.Name}")
-            player_vars.append(player_var)
-
-        # Number of players on team must equal the roster_size
-        # minus one (for a streamer player)
-        model.Add(sum(player_vars) == n_team)
-
-        # Positional constraints
-        pg_target_terms = []
-        sg_target_terms = []
-        sf_target_terms = []
-        pf_target_terms = []
-        c_target_terms = []
-        pg_terms = []
-        sg_terms = []
-        sf_terms = []
-        pf_terms = []
-        c_terms = []
-        g_terms = []
-        f_terms = []
-        for i, player in draft_state_df.iterrows():
-            # Position booleans
-            pos = player["Position"]
-            PG = True if "PG" in pos else False
-            SG = True if "SG" in pos else False
-            SF = True if "SF" in pos else False
-            PF = True if "PF" in pos else False
-            C = True if "C" in pos else False
-            G = True if (PG or SG) else False
-            F = True if (SF or PF) else False
-            if (player["T/A"] == "T") & (player["Drafted"] != 2):
-                pg_target_terms.append(player_vars[i] * PG)
-                sg_target_terms.append(player_vars[i] * SG)
-                sf_target_terms.append(player_vars[i] * SF)
-                pf_target_terms.append(player_vars[i] * PF)
-                c_target_terms.append(player_vars[i] * C)
-            pg_terms.append(player_vars[i] * PG)
-            sg_terms.append(player_vars[i] * SG)
-            sf_terms.append(player_vars[i] * SF)
-            pf_terms.append(player_vars[i] * PF)
-            c_terms.append(player_vars[i] * C)
-            g_terms.append(player_vars[i] * G)
-            f_terms.append(player_vars[i] * F)
-
-        model.Add(sum(pg_terms) >= league_roster["PG"])
-        model.Add(sum(sg_terms) >= league_roster["SG"])
-        model.Add(sum(sf_terms) >= league_roster["SF"])
-        model.Add(sum(pf_terms) >= league_roster["PF"])
-        model.Add(sum(c_terms) >= league_roster["C"] + 2)
-        model.Add(
-            sum(g_terms)
-            >= league_roster["G"] + league_roster["PG"] + league_roster["SG"]
-        )
-        model.Add(
-            sum(f_terms)
-            >= league_roster["F"] + league_roster["PF"] + league_roster["SF"]
-        )
-
-        # Add streamer constraint
-        # streamer_terms = []
-        # for i, player in draft_state_df.iterrows():
-        #     streamer_terms.append(player_vars[i] * (player["Proj$"] <= 1))
-        # model.Add(sum(streamer_terms) == 1)
-
-        # On team and price constraints
-        player_price_terms = []
-        for i, player in draft_state_df.iterrows():
-            if player["Drafted"] == 2:
-                player_price_terms.append(
-                    player_vars[i] * int(round(player["Draft$"], 0))
-                )
-                model.Add(player_vars[i] == 1)
-            elif player["T/A"] == "A":
-                model.Add(player_vars[i] == 0)
-                player_price_terms.append(player_vars[i] * (auction_budget + 1))
-            else:
-                player_price_terms.append(
-                    player_vars[i] * max(int(round(player["Proj$"], 0)), 1)
-                )
-        # Less than the total budget, leaving 1 dollar for the streamer
-        model.Add(sum(player_price_terms) <= auction_budget)
-
-        player_values = []
-        for i, player in draft_state_df.iterrows():
-            player_values.append(player_vars[i] * int(1000 * player.PPW))
-        model.Maximize(sum(player_values))
-
-        solver = cp_model.CpSolver()
-        # solver.parameters.num_search_workers = 12
-        # solver.parameters.log_search_progress = True
-        # solver.parameters.max_time_in_seconds = self.obs_scheme[
-        #     "max_time_in_seconds"
-        # ]
-        status = solver.Solve(model)
-        if status == cp_model.OPTIMAL or status == cp_model.FEASIBLE:
-            players = []
-            for i, player in draft_state_df.iterrows():
-                if solver.Value(player_vars[i]):
-                    players.append(player["Name"])
-            opt_team = create_team_df(
-                players, league_roster, draft_state_df, team_cols, sport
-            )
-        else:
-            opt_team = None
-
-    return opt_team
+# Moved to package.optimization.team_optimizer.find_optimal_team_sat
 
 
-def create_team_df(player_names, league_roster, draft_df, team_cols, sport):
-    if sport == "nfl":
-        relevant_positions = ["QB", "RB", "WR", "TE", "FLEX", "B"]
-        relevant_roster_size = sum([league_roster[pos] for pos in relevant_positions])
-        relevant_positions = ["QB", "RB", "WR", "TE", "FLEX", "B"]
-        league_positions = []
-        league_position_numbers = []
-        for pos, n_required in league_roster.items():
-            if pos not in relevant_positions:
-                continue
-            for npos in range(1, n_required + 1):
-                league_positions.append(pos)
-                league_position_numbers.append(npos)
-        team_dict = {}
-        for col, col_type in team_cols.items():
-            if col == "Position":
-                team_dict[col] = league_positions
-            elif col == "Position rank":
-                team_dict[col] = league_position_numbers
-            elif col_type == str:
-                team_dict[col] = np.repeat("", relevant_roster_size)
-            elif col_type == int:
-                team_dict[col] = np.repeat(0, relevant_roster_size)
-            elif col_type == float:
-                team_dict[col] = np.repeat(0.0, relevant_roster_size)
-
-        team_df = pd.DataFrame.from_dict(team_dict)
-        player_df = copy.deepcopy(draft_df.loc[draft_df["Name"].isin(player_names)])
-        player_df["Position rank"] = player_df.groupby("Position")["PPW"].rank(
-            ascending=False
-        )
-        # Determine flex and bench
-        player_df["FLEX_eligible"] = np.zeros(player_df.shape[0], dtype=bool)
-        flex_positions = ["RB", "WR", "TE"]
-        for pos in flex_positions:
-            flex_mask = (player_df["Position"] == pos) & (
-                player_df["Position rank"] > league_roster[pos]
-            )
-            pos_flex = player_df.loc[flex_mask]
-
-            if not pos_flex.empty:
-                player_df.loc[flex_mask, "FLEX_eligible"] = np.repeat(
-                    True, pos_flex.shape[0]
-                )
-        player_df.loc[player_df["FLEX_eligible"], "FLEX rank"] = player_df.loc[
-            player_df["FLEX_eligible"], "PPW"
-        ].rank(ascending=False)
-        used_players = []
-        for pos, rank in zip(league_positions, league_position_numbers):
-            if pos == "FLEX":
-                player = player_df.loc[(player_df["FLEX rank"] == rank)]
-            else:
-                player = player_df.loc[
-                    (player_df.Position == pos) & (player_df["Position rank"] == rank)
-                ]
-            if not player.empty:
-                for col in team_cols.keys():
-                    if col in ["Position", "Position rank"]:
-                        continue
-                    team_df.loc[
-                        (team_df.Position == pos) & (team_df["Position rank"] == rank),
-                        col,
-                    ] = player[col].values[0]
-                used_players.append(player["Name"].values[0])
-        # Add bench players
-        remaining_players = [name for name in player_names if name not in used_players]
-        if len(remaining_players) > 0:
-            for i, player_name in enumerate(remaining_players):
-                bench_row = len(used_players) + i
-                player = player_df.loc[(player_df.Name == player_name)]
-                for col in team_cols.keys():
-                    if col in ["Position", "Position rank"]:
-                        continue
-                    team_df.loc[bench_row, col] = player[col].values[0]
-    elif sport == "nba":
-        relevant_positions = ["PG", "SG", "SF", "PF", "C", "G", "F", "UTIL", "B"]
-        relevant_roster_size = sum([league_roster[pos] for pos in relevant_positions])
-        league_positions = []
-        league_position_numbers = []
-        for pos, n_required in league_roster.items():
-            if pos not in relevant_positions:
-                continue
-            for npos in range(1, n_required + 1):
-                league_positions.append(pos)
-                league_position_numbers.append(npos)
-        team_dict = {}
-        for col, col_type in team_cols.items():
-            if col == "Position":
-                team_dict[col] = league_positions
-            elif col == "Position rank":
-                team_dict[col] = league_position_numbers
-            elif col_type == str:
-                team_dict[col] = np.repeat("", relevant_roster_size)
-            elif col_type == int:
-                team_dict[col] = np.repeat(0, relevant_roster_size)
-            elif col_type == float:
-                team_dict[col] = np.repeat(0.0, relevant_roster_size)
-
-        team_df = pd.DataFrame.from_dict(team_dict)
-        team_df["Eligible"] = np.repeat(0.0, relevant_roster_size)
-        player_df = copy.deepcopy(draft_df.loc[draft_df["Name"].isin(player_names)])
-        player_df["PPW rank"] = player_df["PPW"].rank(ascending=False)
-        # Rank players into their spots
-        player_df["n_positions"] = np.zeros(player_df.shape[0])
-        for pos in relevant_positions:
-            player_df[f"{pos}_eligible"] = np.zeros(player_df.shape[0])
-        for i, player in player_df.iterrows():
-            # Position booleans
-            pos = player["Position"]
-            PG = True if "PG" in pos else False
-            SG = True if "SG" in pos else False
-            SF = True if "SF" in pos else False
-            PF = True if "PF" in pos else False
-            C = True if "C" in pos else False
-            player_df.at[i, "n_positions"] = PG + SG + SF + PF + C
-            player_df.at[i, "PG_eligible"] = PG
-            player_df.at[i, "SG_eligible"] = SG
-            player_df.at[i, "SF_eligible"] = SF
-            player_df.at[i, "PF_eligible"] = PF
-            player_df.at[i, "C_eligible"] = C
-            player_df.at[i, "G_eligible"] = PG or SG
-            player_df.at[i, "F_eligible"] = SF or PF
-            player_df.at[i, "UTIL_eligible"] = 1
-            player_df.at[i, "B_eligible"] = 1
-        # In the team_df fill in the onesie positions first
-        used_players = []
-        for pos, rank in zip(league_positions, league_position_numbers):
-            remaining_players = [
-                name for name in player_names if name not in used_players
-            ]
-            remaining_player_df = player_df[player_df["Name"].isin(remaining_players)]
-            # Check if there are any players who have eligibility for only the
-            # current position
-            if pos in ["PG", "SG", "SF", "PF", "C"]:
-                eligible_players = remaining_player_df.loc[
-                    (remaining_player_df.n_positions == 1)
-                    & (remaining_player_df[f"{pos}_eligible"] == 1)
-                ]
-                if len(eligible_players) == 0:
-                    eligible_players = remaining_player_df.loc[
-                        (remaining_player_df.n_positions >= 1)
-                        & (remaining_player_df[f"{pos}_eligible"] == 1)
-                    ]
-            elif pos in ["G", "F"]:
-                eligible_players = remaining_player_df.loc[
-                    (remaining_player_df.n_positions == 2)
-                    & (remaining_player_df[f"{pos}_eligible"] == 1)
-                ]
-                if len(eligible_players) == 0:
-                    eligible_players = remaining_player_df.loc[
-                        (remaining_player_df.n_positions >= 2)
-                        & (remaining_player_df[f"{pos}_eligible"] == 1)
-                    ]
-            else:
-                eligible_players = remaining_player_df
-
-            # Highest scoring player with the necessary eligibility
-            if len(eligible_players) > 0:
-                # if not player.empty:
-                player = eligible_players.iloc[0]
-                for col in team_cols.keys():
-                    if col in ["Position", "Position rank"]:
-                        continue
-                    elif col == "Eligible":
-                        team_df.loc[
-                            (team_df.Position == pos)
-                            & (team_df["Position rank"] == rank),
-                            col,
-                        ] = player["Position"]
-                    else:
-                        team_df.loc[
-                            (team_df.Position == pos)
-                            & (team_df["Position rank"] == rank),
-                            col,
-                        ] = player[col]
-                used_players.append(player["Name"])
-
-    return team_df
+# Moved to package.optimization.team_optimizer.create_team_df
