+++ package/optimization/__init__.py
@@ -1,0 +1,1 @@
+# Optimization package
