+++ package/plotting/__init__.py
@@ -1,0 +1,1 @@
+# Plotting package
