+++ pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "auction-draft-gui"
+version = "0.1.0"
+requires-python = ">=3.10"
+
+[tool.ruff]
+line-length = 100
+target-version = "py310"
+select = ["E", "F", "I", "B"]
+ignore = []
+
+[tool.ruff.isort]
+known-first-party = ["package"]
+combine-as-imports = true
